fix linting clippy detected errors

+++ src-core/src/market_data/providers/yahoo_provider.rs
@@ -243,7 +243,7 @@ impl YahooProvider {
         fallback_currency: String,
     ) -> ModelQuote {
         let quote_timestamp: DateTime<Utc> = Utc
-            .timestamp_opt(yahoo_quote.timestamp as i64, 0)
+            .timestamp_opt(yahoo_quote.timestamp, 0)
             .single()
             .unwrap_or_default();
         let now_utc: DateTime<Utc> = Utc::now();
